requirements.txt
arabic-reshaper==3.0.0
asgiref==3.10.0
asn1crypto==1.5.1
attrs==25.4.0
autobahn==24.4.2
Automat==25.4.16
beautifulsoup4==4.14.2
certifi==2025.11.12
cffi==2.0.0
channels==4.3.2
charset-normalizer==3.4.4
constantly==23.10.4
cryptography==46.0.3
cssselect2==0.8.0
daphne==4.2.1
dj-database-url==3.0.1
Django==5.2.8
django-allauth==65.13.0
djangorestframework==3.16.1
freetype-py==2.5.1
html5lib==1.1
hyperlink==21.0.0
idna==3.11
incremental==24.7.2
lxml==6.0.2
mercadopago==2.3.0
oscrypto==1.3.0
pillow==12.0.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.29.0
pycparser==2.23
pyHanko==0.32.0
pyhanko-certvalidator==0.29.0
PyJWT==2.10.1
pyOpenSSL==25.3.0
pypdf==6.4.0
python-bidi==0.6.7
python-decouple==3.8
python-dotenv==1.2.1
PyYAML==6.0.3
reportlab==4.4.5
requests==2.32.5
rlPyCairo==0.4.0
service-identity==24.2.0
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
svglib==1.6.0
tinycss2==1.5.1
tomli==2.3.0
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
webencodings==0.5.1
whitenoise==6.11.0
xhtml2pdf==0.2.17
zope.interface==8.1.1
